% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{met_departments}
\alias{met_departments}
\title{List Met Museum departments}
\usage{
met_departments()
}
\value{
tibble: department_id (integer), display_name (character)
}
\description{
List Met Museum departments
}
