% Generated by build script
\name{met_search}
\alias{met_search}
\title{Search Met Museum artworks}
\usage{
met_search()
}
\arguments{
\item{query}{Search term}
\item{limit}{Max results to fetch details for (default 20)}
}
\value{tibble of matching artworks (IDs and basic info)}
\description{Search Met Museum artworks}
